Implement memory allocation using the Heap class.
Heap now inherits from the new class HeapBase, which contains most of
the size-independent code and just gets a storage area from Heap.
Heap is implemented as a templated class with an array of the
appropriate size and inherits from HeapBase for the memory management
code. It would also be possible to allocate storage dynamically, that
heap class would then also inherit from HeapBase for the memory
management code.

The only thing that is not implemented yet is merging of free blocks
in case of an allocation failure.