Change TypeDescriptor for the DSW garbage collection algorithm.
For the Deutsch-Schorr-Waite mark & sweep collector, the type
descriptor needs to store pointer offsets itself and can't make
use of a vector. This is now accomplished through a custom
placement allocation function and some pointer magic, with the
pointer offsets being dynamically allocated immediately after the
type descriptor itself in a single allocation.