Change TaggedPointer so it can store arbitrary pointers.
Previously TaggedPointer was templated on the type of pointer to be
stored, which doesn't work for this use case since the type changes
during the mark phase.
Now TaggedPointer stores an arbitrary pointer type with the assignment
and get function templated on the stored type. This is not as
convenient but allows the Deutsch-Schorr-Waite marking algorithm to
work.