Fix assertion failures caused by marked objects.
The type descriptor of a marked object cannot be accessed because it
is assumed that the block pointer contains pointer offsets instead of
the type descriptor. In two places the type descriptor would be
accessed before the mark was removed, or after it was set,
respectively, which is fixed now.