Implement HeapObject.
Heap object is a template class that takes as arguments the type
inheriting from it, as well as the type of heap it should allocate
from. A particular type of managed object can therefore only be
allocated on one heap, however there may be many heaps with different
types of objects allocated from each. Pointers across heaps are not
supported.

Class Heap also gets new static methods for allocating and deallocating
memory and implements a singleton pattern making the constructor
private.